Add "topmost" support on OS X/Linux
The overrideredirect Tk flag only works when the window is initially
created on OS X - setting it later with overrideredirect() disables
window controls but doesn't hide them, and changing it back doesn't
re-enable the controls. This probably varies among window managers
on Linux, so I chose to disable overrideredirect() there as well -
the topmost flag is probably more useful for most users, in any case.